PacketNotifyLock
