+++ src/main/java/com/norwood/mcheli/aircraft/MCH_MissileDetector.java
@@ -1,6 +1,6 @@
 package com.norwood.mcheli.aircraft;
 
-import com.norwood.mcheli.networking.packet.MCH_PacketNotifyLock;
+import com.norwood.mcheli.networking.packet.PacketNotifyLock;
 import com.norwood.mcheli.weapon.MCH_EntityBaseBullet;
 import com.norwood.mcheli.wrapper.W_Lib;
 import com.norwood.mcheli.wrapper.W_McClient;
@@ -40,8 +40,8 @@ public class MCH_MissileDetector {
                     if (this.ac.haveFlare() && !this.ac.isDestroyed()) {
                         for (int i = 0; i < 2; i++) {
                             Entity entity = this.ac.getEntityBySeatId(i);
-                            if (entity instanceof EntityPlayerMP) {
-                                MCH_PacketNotifyLock.sendToPlayer((EntityPlayerMP) entity);
+                            if (entity instanceof EntityPlayerMP entityPlayerMP) {
+                                new PacketNotifyLock().sendToPlayer(entityPlayerMP);
                             }
                         }
                     }
